Updated Talk application
I've update the Talk application so that you can now launch in server or client mode. When launched in server mode, a tcp server is instantiated that listens on the localhost:8080 port for any incoming connections while running, if it receives a connection, it can then print out anything that it reads from the connection to the command line. When launched in client mode, the application only currently tries to connect to localhost:8080, and as such is only currently useful for testing. Once a connection is secured, the client can just type into their command line and send, with the server receiving the sent message as a byte slice.