Broke Talk.go
I broke it. The primary issue lies in the fat that the go routines are defined and used in the for loop, leading to unchecked instantiation of new go routines, leading to complete cpu utilization, which leads to a system crash. Those go functions need to be moved, and I need to abstract my communication handling out of the client and server functions and into its own function.