Fixed some parts of Talk.go
I have gone through, pulled the communications handling into its own function called handleComm(), which accepts parameters for a net.Conn connection and a string representing the name of the mode, whether client or server, and uses that for the acknowledgement of the connection. In doing this, I have removed the instantiation of a connection that I was, for some reason, doing in the infinite COMM_LOOP, which was causing its own issues. Still need to fix the fatal system crashing error.