Updated Talk.go
I created a variable to hold a prefix for every message, that prefix being a name like client or server, followed by a colon and a space, "<name>: ", so that it is clear who is communicating what during chat. I also adjusted the formatting to achieve the same goal. The next step is to set up the auto mode, which is meant to first try connecting to a server using provided arguments for hostname or ip address, then just resort to establishing itself as a server if that does not work. The first step of this part is to actually make the program able to accept a hostname or ip address from the arguments.